Implement overloaded not_before() and not_after().
not_before() and not_after() are now functions, instead of strings.

If given a string as a argument they will perform an exact match against all
certificates using ASN1_TIME_print() as before. If given an integer they will
call X509_cmp_time() to compare against all values.

When given an integer not_before() will return true if the timestamp in the
integer is equal to or later than the notBefore attribute of any of the
certificates.

When given an integer not_after() will return true if the timestamp is
"before" (not inclusive) the notAfter attribute of any of the certificates.

There is still a problem with functions in an array of structures if not
scanning a PE file. I still have to debug that.