Parse names out of resource directory entries.
If the Name field has the high bit set the remaining 31 bits are an
offset relative to the start of the resource section. At that offset is
a resource directory string, which contains 2 bytes for the length and a
variable length unicode string.

Teach the resource parser about these by calling a function which will
return a pointer to the resource directory string or NULL if there is no
string or the entire string is not available.

The callback - pe_collect_resources() - takes the pointers to the
strings and calls set_sized_string() on them so they are exposed to
rules.

While I'm here consistently use semicolons in the declare_* lines.