Add section characteristics.
This make it useful to be more specific which sections you want flag.
For example, you can trigger on the entropy of only executable sections.

The list of flags is from section 3.1 of the specification but in the
interest of keeping the list short I stripped out all the reserved or
object file only values.