PE module: Add lower boundary to access bounds check.
On a 32-bit Linux system, I witnessed a crash around
pe_parse_import_descriptor() (pe.c:867) because strndup() wanted to
read from an invalid address that was the result from an integer
overflow and turned out to be less than pe->data.

Since that overflow does not occur on 64-bit systems, they are not
affected.