Remove the "is_loaded" field from module entries to avoid multithreading issues.
When two threads have the same tidx (which is fine if they are operating with different YR_RULES objects) they will collide while using the same bit within the "is_loaded" global bit mask. This may lead to one thread unloading the module loaded by the other. "is_loaded" bit mask is not actually needed, the existence of an object associated to the module in yr_modules_unload_all is enough indicator that the module was loaded.