Add exception handling when scanning memory blocks
Unix: signal handler for SIGBUS + sigsetjmp/siglongjmp
Win32: vectored exception handler + setjmp/longjmp

This should eliminate crashes when accessing the contents of files
that turn out to be unreadable after they have successfully beea
mapped into memory.

This can occur if the mapped file is truncated (Unix) or if the
filesystem is corrupted.