Fix issue causing ungreedy regular expressions resulting in greedy matches
The regular expression /m.*?ssi/ when applied to "mississippi" was matching "mississi" instead of "missi". That's because the chosen atom "ssi" is located twice in the target string causing the regexp being applied backwards two times. The first time it matches "missi" and the second time it matches "mississi". The longer match was replacing the shortest one.

Now the longer match replace the shortest one only if all quantifiers are greedy, if all are ungreedy the shortest match prevail. Mixing greedy and ungreedy quantifiers is forbidden.