Allow tags to be used in conditions.
When a compiler is created a hash table (tags_table) is created which
stores a tag as the key. When a rule is reduced via
yr_parser_reduce_rule_declaration_phase_2() each tag for the rule is
stored in the tags_table. The value is not important as we only care if
the tag exists or not.

There is a new function called yr_parser_store_tag() which will write
the tag identifier to the sz_arena and set the "tag" argument to point
to the newly written data. If the tag identifier is not known - ie: it
wasn't added to tags_table in
yr_parser_reduce_rule_declaration_phase_2() - then an
UNDEFINED_IDENTIFIER is returned.

When an identifier is handled in grammar.y and it is not an object or a
rule then it must be a tag. The tag is stored in sz_arena using
yr_parser_store_tag() and two instructions are emitted:

OP_PUSH 1
OP_PUSH_TAG(tag)

When using the "N of tag" syntax the following instructions are emitted:

OP_PUSH N # Handled by for_expression
OP_PUSH_TAG(tag).

Add the handler for OP_PUSH_TAG instruction, which will get the tag as
an argument, pop the number of rules to match and walk each tag of
matching rules until at least N rules have matched.