Implement imports_ordinal().
In pe_parse_import_descriptor() track the ordinal number used when
importing by ordinal. Store this to be used later in imports_ordinal().

Add an imports_ordinal() which takes a string (DLL name) and integer
(ordinal number). It will then walk the parsed imports looking for a
match.

This allows you to do things like:

import "pe"

rule import_test {
  condition:
    pe.imports("WS2_32.DLL", 3) and
    pe.imports("WS2_32.DLL", "htonl") and
    pe.imports("ntdll.dll") and
    pe.imports("KERNEL32.DLL", "HeapAlloc")
}

Previously you would need to use pe.imports("FOO.DLL", "ord1") to check
for an import by ordinal, but it would only work for those which are not
resolved by ord_lookup(). With this change you can look by either the
resolved name (as is done in the second line above) or by the ordinal
number (as is done in the first line above).