Fix crash in libcrypto when multithreaded.
When running multi-threaded you must properly initalize OpenSSL for
threads. This is done by providing an array of mutexes that is at least
CRYPTO_num_locks(3) long. You then provide a callback using
CRYPTO_set_locking_callback(3) which OpenSSL will call as necessary.
This locking function will either lock or unlock a mutex from the array
based upon a flag provided. You also need to provide a callback for
OpenSSL to use to get the current thread id, which it uses to determine
which lock to operate on.

Failure to do this was causing crashes which would only appear in
multi-threaded runs, and only when doing things involving libcrypto. The
crash is most easily triggered by running a bunch of signed PE binaries
through.

The crash always looked like this:

*** error for object 0x7f983140afc8: incorrect checksum for freed object
- object was probably modified after being freed.

Thread 2 Crashed:
0   libsystem_kernel.dylib          0x00007fff8efde286 __pthread_kill + 10
1   libsystem_c.dylib               0x00007fff874519b3 abort + 129
2   libsystem_malloc.dylib          0x00007fff955c41cb free + 428
3   libcrypto.0.9.8.dylib           0x00007fff92e88e45 CRYPTO_free + 37
4   libcrypto.0.9.8.dylib           0x00007fff92ee0af3 ERR_get_state + 579
5   libcrypto.0.9.8.dylib           0x00007fff92ebf85c ERR_clear_error + 12
6   libcrypto.0.9.8.dylib           0x00007fff92e4e3f8 asn1_d2i_read_bio + 56
7   libcrypto.0.9.8.dylib           0x00007fff92e4e74c ASN1_item_d2i_bio + 44
8   yara                            0x000000010895f2ed pe_parse_certificates + 333
9   yara                            0x0000000108962b49 pe__load + 2233
10  yara                            0x000000010897ab5a yr_modules_load + 266
11  yara                            0x00000001089712ac yr_execute_code + 3564
12  yara                            0x00000001089856f9 yr_rules_scan_mem_blocks + 1161
13  yara                            0x0000000108985af1 yr_rules_scan_file + 113
14  yara                            0x000000010895a042 scanning_thread + 274
15  libsystem_pthread.dylib         0x00007fff8c3e405a _pthread_body + 131
16  libsystem_pthread.dylib         0x00007fff8c3e3fd7 _pthread_start + 176
17  libsystem_pthread.dylib         0x00007fff8c3e13ed thread_start + 13

When the crash happened there would usually be at least one other thread
deep within the OpenSSL internals, usually (but not always) in the same
ASN1_item_d2i_bio() chain.