Bump MAX_PE_IMPORTS to 16384.
Because the imports are parsed and stored for later analysis anytime we
run into the limit, imphash(), imports() and related functions would
return incorrect results when compared to other tools.

This is a compromise to bump it up from 256 to 16384 but it ultimately
doesn't fix the problem that was introduced in 1cc98f7. It seems to be a
conflict of interest between parsing accuracy and memory usage.