Fix build on OS X.
With the latest Xcode on OS X, Apple removed the OpenSSL headers but
left the libraries. This meant the configure script was finding
libcrypto in /usr/lib but building things including OpenSSL headers
would fail.

Since building anything that uses OpenSSL is now broken, the only
reasonable way forward is to use OpenSSL from brew (or something else).
This commit fixes the build to look in the standard brew location for
OpenSSL for OS X.

For more information see this link:
http://lists.apple.com/archives/macnetworkprog/2015/Jun/msg00025.html

Fixes #351.