Add support for CLI parsing.
This commit adds support for parsing the COM Runtime descriptor
directory entry, which is only the start of the rabbit hole. ;)

Ultimately this leads us to parsing the CLI MetaData streams, including
GUIDs and a convoluted table structure containing various information
relevant to the CLI.

In particular, here are the things we parse out:

+  declare_string("dotnet_version");
+  declare_string("dotnet_module_name");
+  begin_struct_array("dotnet_streams");
+    declare_string("name");
+    declare_integer("offset");
+    declare_integer("size");
+  end_struct_array("dotnet_streams");
+  declare_integer("number_of_dotnet_streams");
+  declare_string_array("dotnet_guids");
+  begin_struct_array("dotnet_resources");
+    declare_integer("offset");
+    declare_integer("length");
+    declare_string("name");
+  end_struct_array("dotnet_resources");
+  declare_integer("number_of_dotnet_resources");
+  begin_struct("dotnet_assembly");
+    begin_struct("version");
+      declare_integer("major");
+      declare_integer("minor");
+      declare_integer("build_number");
+      declare_integer("revision_number");
+    end_struct("version");
+    declare_string("name");
+    declare_string("culture");
+  end_struct("dotnet_assembly");
+  declare_string_array("dotnet_modulerefs");
+  declare_integer("number_of_dotnet_modulerefs");

Unfortunately there are a lot of manual table walking because the tables
contain variable size columns, which depend upon the number of rows in
other tables. This makes the table walking code more messy than I would
like.

And since I'm here and noticed it was unused, remove the unused
RICH_DATA structure.