Use memmem(3) instead of strnstr(3).
strnstr(3) is not portable and rather than include it just for this one case
fake it using memmem(3). If necessary we can add our own implementation of
memmem(3) and do all the necessary autoconf stuff to detect.

While here, include stdarg.h because it is necessary to suppress a warning
with gcc about using va_start(3) and friends.