Change include path while compiling in Mac OS X
The current path was set in 9b6c755e0dce42920058a181adf2da1871588758 to fix an issue with missing OpenSSL headers in Mac OS X. I'm changing the path to /opt/local/include which is the include path used by MacPorts. I'm also removing the -L argument because it was ignored by clang.