Fix RVA calculation corner cases.
When the PointerToRawData is less than 0x200 it is rounded down to 0x00. This
causes entry point calculation to be incorrect in some cases. For example, these
two issues highlight the problem:

https://github.com/plusvic/yara/issues/399
https://github.com/plusvic/yara/issues/373

Details of this are in
https://code.google.com/archive/p/corkami/wikis/PE.wiki#PointerToRawData

I tested this against 293 PE files and found one difference in entry point
calculation, which upon inspection turned out to be another case of this bug.

Fixes: #399
Fixes: #373