Fix issue #516.
When attempting to import "\x00" there is a crash due to an integer underflow.
The "\x00" is passed to yr_parser_reduce_import() as a SIZED_STRING, with a
c_string member pointing to "\x00" and a length of 1. Eventually
yr_hash_table_lookup() is given the c_string pointer, which then calls strlen()
on it and passes the length (0) into yr_hash_table_lookup_raw_key(). So at this
point yr_hash_table_lookup_raw_key() is given a pointer to a string containing
"\x00" and a length of 0, which it then passes to hash() which ultimately
subtracts 1 from the 0 length and then looks that value up in the byte_to_int32
array, which is obviously bogus.

This fixes the problem by checking for a key_length of 0 in
yr_hash_table_lookup_raw_key() before attempting to hash it and returns NULL.
The NULL value is now checked for in yr_parser_reduce_import() and a compiler
error is now thrown.

I did a quick check for other places where a similar problem may pop up,
modules are especially concerning, but I haven't found any that don't look right
to me.