Revert previous commit and fix it properly.
Turns out checking for the underflow and returning NULL there is not a good idea
because NULL is used to indicate that no such object exists in the hash table,
and we don't want to hijack that to mean error. This was ultimately what was
causing a valid import to be unrecognized, which is why the tests were failing.

Instead of checking for the underflow I'm now just checking if the sized string
length is zero or if the strlen() of the c_string is zero and erroring out
early.

While here, add a test for this.