Don't unmask signals when setting up exception handler (Unix) (#546)
* YR_TRYCATCH: Remove duplicate initialization of oldmask

* Add test for handling exceptions

With

    #define YR_TRYCATCH(_try_clause_, _catch_clause_) {_try_clause_}

this will cause a bus error on Linux -- as it should.

* tests: Add assert_*_rule_blob_size, make count_matches usable from test code

* test-exception: Block and send SIGUSR1 before 2nd yr_rules_scan_mem call

This seems to reproduce something similar to the symptom described in

$ ./test-exception
Scanning for "aaaa"...
err = 4, matches = 0
Sending blocked SIGUSR1 to ourselves...
Scanning for {00 00 00 00}...
User defined signal 1

* exception (UNIX): Remove unneeded pthread_sigmask() calls

sigsetjmp() already saves the signal mask for us.

Also, setting the signal mask using the (empty) act.sa_mask was wrong
and led to the behavior described in #544.

We still want delivery of all signals to be blocked during execution of
the signal handler.

* Document test-exception

cherry-picked from:

- 8b2b6b8f82b8a8a8830bae122474b39ac2814fc6
- 0b3d9a93276d3364b2a3561995d27e8b4a137134, "Minor re-styling"