Add APIs for defining contexts for local external variables
This set of APIs allows for thread-safe manipulation of external
variables. Applications can allocate several contexts for external
variables which are initialized as copies of the default external
variable set (yr_rules_allocate_local_variable_context(),
yr_rules_free_local_variable_context()). Those contexts are identified
using a positive numeric value (uint16_t).

These contexts can then be modified and safely used by multiple threads
for scanning, provided that the application does not modify a context
which it currently uses for scanning. For modifications of the external
variables, a set of yr_rules_define_local_*_variable() functions are
added that accept an added context argument.

The context is passed to the existing yr_rules_scan_*() functions as
part of the flags field, using the SCAN_EXTERNAL_VAR_CONTEXT(ctx) macro.
The default case context=0 means that the default external variables are
used: This ensures API compatibility for programs that do not use the
new feature.

The layout for compiled rulesets as used in yr_rules_{load,save}*() is
not changed.