Handle malformed streams better. (#653)
When parsing the #~ stream it will require references to the #US and #Blob
streams, so we need to make sure those are not NULL before attempting to parse
the stream. It is possible to only do these checks in the place we need them but
I'd rather declare the whole thing as malformed once and early.

Fixes part of #650.