Remove unnecessary null check.
Coverity raised the following warning:

```
check_after_deref: Null-checking context->proc_info suggests that it may be null, but it has already been dereferenced on all paths leading to the check.
```

That's because _yr_process_detach dereferences context->proc_info without any checking, so either _yr_process_detach is right about the assumption that context->proc_info can not be null, or yr_process_close_iterator is wrong about the assumption that it can be null.