modules/pe: Fixed possible invalid memory read when parsing certificates in PE
Attribute Length in WIN_CERTIFICATE structure is length of bCertificate buffer,
not the length of the whole structure. Every time we want to work with the Certificate buffer,
we need to subtract size of the header. If the security directory is the
very last thing in the PE file, we can possibly end up with invalid pointer.