Open files in default include callback with binary mode callback on Windows (fixes #809)
Default include callback obtains the file size of included file
in order to allocate sufficiently large buffer. It then compares
the file size from _filelength() with the return value of read()
call. However, _filelength() most probably just read the metadata
of the file and returns the amount of bytes in the file. In contrary,
read() on Windows transforms CRLF line endinges to LF line endinges
when file is open in text mode and thus read() returns different value
than the actual file size. Opening the file in binary mode should
solve this problem.