Implement the scanner API.
A new set of functions yr_scanner_* has been added to the API. The purpose of these functions is providing a way for using the same YR_RULES object with different sets of values for external variables. With the existing yr_rules_scan_* functions all threads sharing the YR_RULES where forced to use the same set of values for externals variables, as the values where associated to the YR_RULES object. With the new API each thread will be able to create its own YR_SCANNER object to be used with the same underlying YR_RULES object. Each YR_SCANNER can have a different value for the same external variable. The existing API will work as before.