Don't access directory entries past SizeOfOptionalHeader. (#802)
Attempt to make pe_get_directory_entry() a little more resilent to malformed PE
files.

As discussed in a comment in the code,
7ff1bf680c80fd73c0b35084904848b3705480ddeb6d0eff62180bd14cd18570 is an
interesting sample which has NumberOfRvaAndSizes set to 11 but does in fact have
more than 11 entries in the DataDirectory array. We happened to work for this
sample because we were not checking NumberOfRvaAndSizes before accessing the
DataDirectory array. While the docs say the NumberOfRvaAndSizes is to be
checked. If we were to enforce this strict check to ensure safety we would be
incorrectly parsing what is an otherwise valid .NET binary.

Instead of being overly strict here I'm attempting to make sure the entry in the
DataDirectory array is at least within the bounds of the OptionalHeader. If it
isn't the function now returns NULL, instead of whatever happened to be in
memory at that address. Since the function now returns NULL I have to fix every
caller to check for it and return accordingly.

While I'm here, I might as well remove a redundant prototype for
pe_get_directory_entry in pe_utils.h and add in NumberOfRvaAndSizes because it's
useful to catch packers which alter this value.