Fix issue when compiling YARA with MAX_ATOM_SIZE larger than 4.
The size of the buffer used by _yr_atoms_case_insensitive was correct for MAX_ATOM_SIZE == 4 but the formula for calculating it was incorrect. That caused a buffer-overflow when increasing MAX_ATOM_SIZE. This  fixes the formula for calculating the buffer size and reduces memory consumption by using uint8_t instead of int for the atom lengths stored in the buffer.