Corrected early stoppage of Imported DLLs parsing.
When the program encounters an IMAGE_IMPORT_DESCRIPTOR whose DLL name is not valid, it stops parsing the current descriptor and the subsequent ones. It should be important to prosecute the parsing, at least of the next descriptors, in order not to miss any imported DLL. A malware author could "obfuscate" the entire set of Imported DLLs by putting, as first DLL, one with an invalid Name.