Corrected early stoppage of Imported DLLs parsing. (#896)
When the program encounters an IMAGE_IMPORT_DESCRIPTOR whose DLL name is not valid, it stops parsing the current descriptor and the subsequent ones. It's important to continue the parsing in order not to miss any imported DLL. A malware author could "obfuscate" the entire set of imported DLLs by putting, as the first DLL, one with an invalid name.