Add parentheses to yr_min and yr_max parameters
If the parameters given to yr_min and yr_max macros use operations which
priority is lower than the one of "<", the current definitions may cause
unexpected results. Adding parentheses makes sure this can never happen.

For information, the current implementations of yr_min and yr_max
evaluate twice their parameters. This could be an issue when a parameter
is a function call. Nevertheless MSVC does not provide a way to perform
something similar this gcc-specific code:

    #define yr_min(x, y) ({ \
        typeof(x) _min1 = (x); \
        typeof(y) _min2 = (y); \
        _min1 < _min2 ? _min1 : _min2; \
        })

Therefore, do not change the implementation of yr_min and yr_max this
way.