Fix an OOM condition in dex module.
When scanning certain files the dex module will OOM. The file seems to be
obfuscated somehow and it causes the parser to misbehave. I haven't completely
narrowed down what is going on, but this change at least allows the parser to
complete without excessive memory usage and also quickly.

This is easily triggerable with
c8a3172a42cd9270db562b8fd0fddc15d35511483818bfc01c508fdb39f01635.

This change does two things:

1. Move the check for name_idx on both fields and methods to earlier in the
function. If the name_idx is UNDEFINED return early and indicate to the caller
that it should stop parsing that piece because it's not going to work.

2. The parser now support abandoning parsing certain pieces early if the
corresponding function (load_encoded_field or load_encoded_method) returns 0.
The parser would otherwise spend A LOT of time parsing files that are obviously
"wrong."

It does cause some method names (and possibly other things) to be incorrectly
parsed. For example, there's this output now when running the module on the
above sample:

[1]
		class_name = "Landroid/animation/AnimatorSet;"
		name = "le\x00\x11setExistBitmapImg\x00\x0bsetExpanded\x00\x09setExport\x00\x0dsetExportPath\x00\x0fsetExtActionBar\x00\x13setExtendedProperty"
		proto = "LJ"
		direct = 0
		virtual = 1
		method_idx_diff = 17
		access_flags = 1
		code_off = 17
		code_item
				registers_size = 52336
				ins_size = 46768
				outs_size = 24434
				tries_size = 46705
				debug_info_off = 1781676661
				insns_size = 74018772
				insns = UNDEFINED
				padding = UNDEFINED
				tries
				handlers