Optimize a little further.
Add a OP_SET_M which takes an argument (the memory slot) and will set it equal
to the value on the stack, but does not alter the state of the stack.

With this instruction I reduced the sequence of:

OP_CLEAR_M 4
OP_ADD_M 4
OP_PUSH_M 4

down to just:

OP_SET_M 4

I also moved the (now new) OP_SET_M into the _FOR_ for_expression _IDENTIFIER_ _IN_
block, so we don't add these unnecessary instructions for all loops. They were
harmless from a execution perspective but completely unnecessary for
performance.

Lastly, go ahead and replace the sequence:

OP_POP_M 4
OP_PUSH_M 4

with:

OP_SET_M 4

This is the sequence used when storing the last result of the boolean_expression
in "all" loops.