Optimize integer range loops. (#1031)
Optimize integer range loops. If a loop can not possible evaluate to true, stop processing it. For example:

for all i in (0..100): (false)
for any i in (0..100): (true)

Both of these loops will stop iterating after the first time through.

Also, loops which use a primary expression are also handled (they are just a special case of "any" loops).