Change export parsing.
Change export parsing so that we now parse more data than just the name. The new
structures exposed look like this:

pe.dll_name: The DLL name (from the export directory).
pe.export_timestamp: The timestamp (from the export directory).
pe.export_details[].offset: Where the function starts (NOT the RVA).
pe.export_details[].name: The exported name (if it exists).
pe.export_details[].forward_name: The forwarder name (if it exists).
pe.export_details[].ordinal: The ordinal (with the ordinal base added to it).

Prior to this change we were storing the ordinal without adding in the ordinal
base, which is not how most tools do it. I checked IDA and pestudio and both
display the export ordinal after the ordinal base has been added to it. This
makes sense because that is the actual ordinal you need to use to import the
function.

I've also adjusted the "pe.exports()" functions so they now walk the collected
export data exposed in the module. This keeps backwards compatability for rules
that use "pe.exports()", but you can get the same behavior by using:

for any i in (0..pe.number_of_exports - 1):
  (pe.export_details[i].name == "AXS_ERS")

I'm considering deprecating the pe.exports() functions since they are
technically no longer necessary, but have opted to not do that in this diff.

I've also chosen to add a sized_string_cmp_nocase() since pe.exports() used to
do a case insensitive comparison.

One thing to remember with this change is that since we now respect the ordinal
base, any rule using pe.exports() with an ordinal is going to behave
differently.  For example, looking at mscoree.dll from a Windows 10 machine the
ordinal base is 17 so an old rule that did pe.exports(1) would match before this
change but not after it since now 17 is the lowest possible ordinal value we
will store.