Private strings (#1096)
* Add private modifier for strings.

This commit adds a private modifier to the grammar, which can be used on any
string. When this modifier is used the string is treated as normal except it is
never reported back in match objects. It is still searched for normally and
still used normally in condition evaluation but it will never be seen in the
match callback.

This is a useful feature when you have strings that you want to use but they
match a lot of times and you don't want them polluting your log results. The
best example that demonstrates this is the technique described at [1]. In this
example the local file header magic is going to match a lot of times, and it is
only needed to look through the local file header for filenames at offsets
relative to the matches. We don't really care where the offsets of the local
file header are, just where the matching filenames are.

If you are logging where all the matches are for every string and there are
10,000 local file headers that's a lot of log data being generated which is not
helpful at all.

[1]: https://blog.korelogic.com/blog/2015/08/19/yara-ctf-2

* Optimize VM execution code for private strings.

Rather than checking the matches list and only if none are found then check for
private matches, it's better to check only the appropriate list after using
STRING_IS_PRIVATE.

* Add private strings in various tests.

This should cover all the cases of operators on strings that also need to check
for private strings.

* Add docs for private strings.

* Fixup documentation.

Private strings apply to all string types, not just text strings.

* Fix private modifier for chained strings. Add a test case for it too.