Fix chained strings (#1107)
* Add test case.

* Fix issue with regexps and hex strings that were not splitted correctly.

The function yr_re_ast_split_at_chaining_point was not splitting correctly regexps and hex strings that contained more than one large jump (a.k.a gap). For example, { 01 02 03 [0-1000] 04 05 06 [500-2000] 07 08 09 } was not splitted in three parts as expected, but only in two parts: { 01 02 03 } and { 04 05 06 [500-2000] 07 08 09 } . This didn't affected the correctness of the matches, but it caused ERROR_TOO_MANY_RE_FIBERS in some cases while trying to match { 04 05 06 [500-2000] 07 08 09 } because of the large jump in the pattern. The occurrence of the error depended on the data being scanned.

This commit fixes the aforementioned issue, improves the documentation of related functions and simplifies some portions of the code.

* Avoid putting the same string multiple times in matching_strings_arena.

This fixes a small issue introduced in #1096. The same string could be put more than once in the matching_strings_arena, this doesn't causes any major error nor affect the correctness of the matching, but it could have a negative impact on performance.

* Add more comments.

* Improve documentation for _yr_scan_verify_chained_string_match.