Fixes for xor modifier. (#1113)
* Fixes for xor modifier.

This commit makes the xor modifier no longer include "plaintext" versions of the
strings. In particular if you have:

$a = "This program cannot" xor

Then it will no longer match on the plaintext version of the string. It will
only match on the xor versions. If you want it to match on plaintext or xor then
you can use:

$a = "This program cannot" xor ascii

The same is true for wide.

* Make 'xor nocase' a compiler error.

* Add newline between ascii and wide strings.