Iterators (#1141)
This introduces a change in the way in which "for-in" loops work by introducing the concept of iterators. These kind of loops now accept an iterator that will return a sequence of values. Besides the integer ranges (N..M) and integer enumerations (X,Y,Z), this kind of loops now accepts other kinds of iterables, like arrays. This allows conditions like the following one:

for any section in pe.sections: ( section.name == "foo" )

Until now the same expression was written as:

for any i in (0..pe.number_of_sections - 1): ( pe.sections[i].name == "foo" )

The new syntax is more legible and opens door for more powerful features in the future. Backward compatibility is maintained.