Fix build by removing PKG_CHECK_MODULES.
I'm not sure if this is a proper fix or not but the build is failing for me
with:

./configure: line 14733: syntax error near unexpected token `PROTOBUF_C,'
./configure: line 14733: `    PKG_CHECK_MODULES(PROTOBUF_C, libprotobuf-c >= 1.0.0)'

As far as I can tell these were left in by mistake so they should be removed?
Once I remove the PKG_CHECK_MODULES and regenerate configure script the build is
working again.