Fix trailing byte in some dotnet user strings. (#1207)
There is always a trailing byte added into user strings and blob entries. This
byte is 0x00 or 0x01 and is used to indicate if any UTF16 character requires
handling beyond the normal 8-bit encoding sets. See ECMA-335 II.24.2.4 for
exact details, which are not relevant at all to our parsing. This commit just
ignores that trailing byte.

Add a test for this too. It's worth noting that this does break anyone who is
looking for specific user strings but given that this module is not on by
default I'm willing to break those people. :(

Noticed by: Robert Simmons