Fix memory leaks and different alphabets.
Most of these memory leaks were silly and were because I was forgetting to
free the encoded string or the nodes upon successful creation of the AST.

There were some leaks when the compiler would error in the case of incorrect
alphabet size.

Lastly, I realized that if you specify two different alphabets in your
modifier the last one used would apply to the other base64 modifier, which is
not intuitive behavior. Fix it by making different alphabets a compiler error.
Add a test case for this while I'm here.